refactor: update Arch package management logic in scripts and config - Reorganize pacman checks in update script to provide informative messages - Conditionalize Arch maintenance abbreviations in fish config based on pacman availability - Enhance package listing function to ensure compatibility with Arch-based systems

+++ fish/.config/fish/config.fish
@@ -32,16 +32,21 @@ function fish_greeting
 end
 
 # 6. Abbreviations
-# --- Arch Maintenance ---
-abbr pacin 'sudo pacman -S'
-abbr pacinn 'sudo pacman -S --needed'
-abbr pacrem 'sudo pacman -Rns'
-abbr orphans 'pacman -Qdtq | sudo pacman -Rns -'
-abbr mirrors 'sudo rate-mirrors --protocol https --allow-root --save /etc/pacman.d/mirrorlist --disable-comments-in-file arch'
-abbr mirrors-aur 'sudo rate-mirrors --disable-comments-in-file --protocol=https --allow-root --save /etc/pacman.d/chaotic-mirrorlist chaotic-aur'
+if type -q pacman
+    # --- Arch Maintenance ---
+    abbr pacin 'sudo pacman -S'
+    abbr pacinn 'sudo pacman -S --needed'
+    abbr pacrem 'sudo pacman -Rns'
+    abbr orphans 'pacman -Qdtq | sudo pacman -Rns -'
+    abbr mirrors 'sudo rate-mirrors --protocol https --allow-root --save /etc/pacman.d/mirrorlist --disable-comments-in-file arch'
+    abbr mirrors-aur 'sudo rate-mirrors --disable-comments-in-file --protocol=https --allow-root --save /etc/pacman.d/chaotic-mirrorlist chaotic-aur'
 
-abbr drop-cache 'sudo paccache -rk3; and sudo pacman -Sc --noconfirm'
-abbr fixpacman 'sudo rm /var/lib/pacman/db.lck'
+    abbr drop-cache 'sudo paccache -rk3; and sudo pacman -Sc --noconfirm'
+    abbr fixpacman 'sudo rm /var/lib/pacman/db.lck'
+
+    abbr big "expac -H M '%m\t%n' | sort -h | nl"
+    abbr rip "expac --timefmt='%Y-%m-%d %T' '%l\t%n %v' | sort | tail -200 | nl"
+end
 
 # --- Dev Stack ---
 abbr art 'php artisan'
@@ -61,5 +66,3 @@ alias fastfetchy 'fastfetch -c $HOME/.config/fastfetchTheme.jsonc'
 abbr tb 'nc termbin.com 9999'
 abbr errors 'journalctl -p 3 -xb'
 abbr hw 'hwinfo --short'
-abbr big "expac -H M '%m\t%n' | sort -h | nl"
-abbr rip "expac --timefmt='%Y-%m-%d %T' '%l\t%n %v' | sort | tail -200 | nl"
